+++ client/src/mivaCommands.ts
@@ -1,5 +1,5 @@
 import patterns from './util/patterns';
-import { TextEditor, TextEditorEdit, Range, commands, env, window, workspace, Uri } from 'vscode';
+import { TextEditor, TextEditorEdit, Range, commands, env, window, workspace, Uri, languages, CharacterPair } from 'vscode';
 
 const boundryAmount = 200;
 
@@ -207,28 +207,26 @@ const convertToVariableCommand = commands.registerTextEditorCommand( 'mivaIde.MV
 
 });
 
-const insertHtmlComment = commands.registerTextEditorCommand( 'mivaIde.toggleHtmlComment', ( textEditor: TextEditor, edit: TextEditorEdit ) => {
+const insertHtmlComment = commands.registerTextEditorCommand( 'mivaIde.toggleHtmlComment', ( textEditor: TextEditor ) => {
 
-	// exit if not MVT or MV
-	if ( textEditor.document.languageId !== 'mvt' && textEditor.document.languageId !== 'mv' ) {
-		commands.executeCommand( 'editor.action.blockComment' );
-		return;
+	const languageId = textEditor.document.languageId;
+
+	if ( languageId === 'mvt' || languageId === 'mv' ) {
+		
+		languages.setLanguageConfiguration( languageId, { comments: { blockComment: [ '<!--', '-->' ] } } );
+
+		commands.executeCommand( 'editor.action.blockComment' ).then(() => {
+
+			languages.setLanguageConfiguration( languageId, { comments: { blockComment: ( languageId == 'mv' ) ? [ '<MvCOMMENT>', '<MvCOMMENT>' ] : [ '<mvt:comment>', '<mvt:comment>' ] } } );
+
+		});		
+
 	}
+	else {
 
-	textEditor.selections.forEach(( selection ) => {
+		commands.executeCommand( 'editor.action.blockComment' );
 
-		// let range = new Range( selection.start, selection.end );
-		// let text = textEditor.document.getText( range );
-		// let cleanedText = text.replace( /(<!--\s*)|(\s*-->)/g, '' );
-		// let replacement = ( cleanedText.length !== text.length ) ? cleanedText : `<!-- ${ text } -->`;
-
-		// edit.replace( range, replacement );
-
-		let text = textEditor.document.getText( new Range( selection.start.line, Math.max(selection.start.character - 5, 0), selection.end.line, (selection.end.character + 4) ) );
-
-		console.log( 'test', text );
-
-	});
+	}
 
 });